Switch to python version of build_and_copy_bin

+++ scripts/build.py
@@ -127,21 +127,22 @@ def find_flash_size(dts_filename):
 
     return flash_name, flash_size
 
-def build_and_copy_bin(zephyr_platform, sample_path, args, sample_name):
+def build_and_copy_bin(zephyr_platform, sample_path, args, sample_name, env):
     zephyr_sample_name = f"{zephyr_platform}-{sample_name}"
-    os.makedirs(f"artifacts/{zephyr_sample_name}")
+    os.makedirs(f"artifacts/{zephyr_sample_name}", exist_ok=True)
     previous_dir = os.getcwd()
     os.chdir(zephyr_path)
     build_path = f"build.{zephyr_platform}.{sample_name}"
-    shutil.rmtree(build_path)
+    if os.path.isdir(build_path):
+        shutil.rmtree(build_path)
     log_path = f"../../artifacts/{zephyr_sample_name}/{zephyr_sample_name}-zephyr.log"
 
-    run_west_cmd(f"west spdx --init -d {build_path}", f"tee -a {log_path}")
-    run_west_cmd(f"west build --pristine -b {zephyr_platform} -d {build_path} {sample_path} {args}", f"tee -a {log_path}")
-    run_west_cmd(f"west spdx -d {build_path}", f"tee -a {log_path}")
+    west_output = run_west_cmd(f"west spdx --init -d {build_path}", env, log_path)
+    west_output += run_west_cmd(f"west build --pristine -b {zephyr_platform} -d {build_path} {sample_path} {args}", env, log_path)
+    west_output += run_west_cmd(f"west spdx -d {build_path}", env, log_path)
 
     os.chdir(previous_dir)
-    file_list=["zephyr/zephyr.elf" "zephyr/zephyr.dts" "zephyr/.config" "spdx/app.spdx" "spdx/build.spdx" "spdx/zephyr.spdx"]
+    file_list=["zephyr/zephyr.elf", "zephyr/zephyr.dts", "zephyr/.config", "spdx/app.spdx", "spdx/build.spdx", "spdx/zephyr.spdx"]
 
     for file_name in file_list:
         file_path = f"{zephyr_path}/{build_path}/{file_name}"
@@ -157,12 +158,15 @@ def build_and_copy_bin(zephyr_platform, sample_path, args, sample_name):
                     shutil.copyfile(file_path, f"artifacts/{zephyr_platform}.dts")
             if file_name == file_list[2]:
                 shutil.copyfile(file_path, f"artifacts/{zephyr_sample_name}/{zephyr_sample_name}-config")
-    shutil.rmtree(f"{zephyr_path}/{build_path}")
+    if os.path.isdir(build_path):
+        shutil.rmtree(f"{zephyr_path}/{build_path}")
+    return west_output
 
-def run_west_cmd(cmd, pipe_cmd):
-    ps = subprocess.Popen((cmd.split(" ")), stdout=subprocess.PIPE)
-    subprocess.check_output(pipe_cmd.split(" "), stdin=ps.stdout)
-    ps.wait()
+def run_west_cmd(cmd, env, log_file):
+    output = subprocess.check_output((cmd.split(" ")), env=env).decode()
+    with open(log_file, 'a') as file:
+        file.write(output)
+    return output
 
 def build_sample(zephyr_platform, sample_name, sample_path, sample_args, toolchain, download_artifacts, skip_not_built):
     if download_artifacts:
@@ -191,34 +195,33 @@ def build_sample(zephyr_platform, sample_name, sample_path, sample_args, toolcha
         return
     print(f"Building for {bold(zephyr_platform)}, sample: {bold(sample_name)} with args: {bold(sample_args)} using {bold(toolchain)} toolchain.")
     args = f'-- {sample_args}' if sample_args != '' else ''
-    process = subprocess.run(["./scripts/build_and_copy_bin.sh", zephyr_platform, sample_path, args, sample_name], stdout=subprocess.PIPE, env=env)
-
-    # try increasing flash size if the sample doesn't fit in it 
+    west_output = build_and_copy_bin(zephyr_platform, sample_path, args, sample_name, env)
+    # try increasing flash size if the sample doesn't fit in it
     dts_filename = artifacts_dict['dts'].format(board_name=zephyr_platform, sample_name=sample_name)
-    if process.returncode:
-        m = re.search(r"region `FLASH' overflowed by (\d+) bytes", process.stdout.decode())
-        if m is not None and os.path.exists(dts_filename):
-            shutil.copy2(dts_filename, dts_filename + '.orig')
-            flash_increase = math.ceil(int(m.group(1)) / 1024) * 1024
-            flash_name, flash_size = find_flash_size(dts_filename)
-            if len(flash_size) >= 2:
-                flash_base, flash_size = flash_size[-2:]
-                flash_size = int(flash_size, 16)
-                flash_size += flash_increase
+    m = re.search(r"region `FLASH' overflowed by (\d+) bytes", west_output)
+    if m is not None and os.path.exists(dts_filename):
+        shutil.copy2(dts_filename, dts_filename + '.orig')
+        flash_increase = math.ceil(int(m.group(1)) / 1024) * 1024
+        flash_name, flash_size = find_flash_size(dts_filename)
+        if len(flash_size) >= 2:
+            flash_base, flash_size = flash_size[-2:]
+            flash_size = int(flash_size, 16)
+            flash_size += flash_increase
 
-                with tempfile.NamedTemporaryFile(mode='w', encoding='utf-8') as f:
-                    f.write(dts_flash_template.render(
-                        flash_name=flash_name,
-                        reg_base=flash_base,
-                        reg_size=flash_size
-                    ))
-                    f.flush()
-                    overlay_path = f.name
+            with tempfile.NamedTemporaryFile(mode='w', encoding='utf-8') as f:
+                f.write(dts_flash_template.render(
+                    flash_name=flash_name,
+                    reg_base=flash_base,
+                    reg_size=flash_size
+                ))
+                f.flush()
+                overlay_path = f.name
 
-                    # build again, this time with bigger flash size
-                    overlay_args = f'-DDTC_OVERLAY_FILE={overlay_path}'
-                    args = f'-- {sample_args} {overlay_args}'
-                    process = subprocess.run(["./scripts/build_and_copy_bin.sh", zephyr_platform, sample_path, args, sample_name], stdout=subprocess.PIPE, env=env)
+                # build again, this time with bigger flash size
+                overlay_args = f'-DDTC_OVERLAY_FILE={overlay_path}'
+                args = f'-- {sample_args} {overlay_args}'
+                build_and_copy_bin(zephyr_platform, sample_path, args, sample_name, env)
+
 
 def get_board_yaml_path(board_name, board_path):
     board_yaml = f'{zephyr_path}/{board_path}/{board_name}.yaml'
